requirement.txt
pyserial
pyautogui
matplotlib
pygame